# HDDS-12341. Share cluster in client tests