## What changes were proposed in this pull request?

Speed up some client integration tests by using shared cluster.

Also, fix intermittent failure caused by background key deletion (HDDS-12465), because it become more frequent with these new tests added.

https://issues.apache.org/jira/browse/HDDS-12341

## How was this patch tested?

Before:

```
Tests run: 7, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 66.09 s -- in org.apache.hadoop.ozone.client.rpc.TestCloseContainerHandlingByClient
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 66.91 s -- in org.apache.hadoop.ozone.client.rpc.TestContainerStateMachineStream
Tests run: 29, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 43.65 s -- in org.apache.hadoop.ozone.client.rpc.TestOzoneClientMultipartUploadWithFSO
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 38.03 s -- in org.apache.hadoop.ozone.client.rpc.TestOzoneRpcClientWithKeyLatestVersion
```

After:

```
Tests run: 7, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 27.18 s -- in org.apache.ozone.test.NonHATests$CloseContainerHandlingByClient
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.233 s -- in org.apache.ozone.test.NonHATests$ContainerStateMachineStream
Tests run: 29, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 2.208 s -- in org.apache.ozone.test.NonHATests$OzoneClientMultipartUploadWithFSO
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 1.539 s -- in org.apache.ozone.test.NonHATests$OzoneRpcClientWithKeyLatestVersion
```

CI:
https://github.com/adoroszlai/ozone/actions/runs/13979021418